+++ includes/class-wp-php-console.php
@@ -2,59 +2,72 @@
 /**
  * WP PHP Console Plugin Core Class
  *
- * @link       https://github.com/nekojira/wp-php-console
- * @since      1.0.0
- *
- * @package    WP_PHP_Console
+ * @link    https://github.com/nekojira/wp-php-console
+ * @since   1.0.0
+ * @package WP_PHP_Console
  */
+namespace WP_PHP_Console;
+
+use PhpConsole;
 
 if ( ! defined( 'WPINC' ) ) {
 	exit; // Exit if accessed directly
 }
 
 /**
- * WP PHP Console.
+ * WP PHP Console main class.
  *
- * @since      1.0.0
- * @package    WP_PHP_Console
+ * @since   1.0.0
+ * @package WP_PHP_Console
  */
-class WP_PHP_Console {
+class Plugin {
+
+	/**
+	 * The plugin name.
+	 *
+	 * @since  1.4.0
+	 * @access public
+	 * @var    string $plugin_name The name of this plugin.
+	 */
+	public $plugin_name = '';
 
 	/**
 	 * The unique identifier of this plugin.
 	 *
-	 * @since    1.0.0
-	 * @access   protected
-	 * @var      string    $plugin_name    The string used to uniquely identify this plugin.
+	 * @since  1.0.0
+	 * @access public
+	 * @var    string $plugin_slug The string used to uniquely identify this plugin.
 	 */
-	protected $plugin_name = '';
+	public $plugin_slug = '';
 
 	/**
 	 * The current version of the plugin.
 	 *
-	 * @since    1.0.0
-	 * @access   protected
-	 * @var      string    $version    The current version of the plugin.
+	 * @since  1.0.0
+	 * @access public
+	 * @var    string $version The current version of the plugin.
 	 */
-	protected $version = '';
+	public $version = '';
 
 	/**
-	 * Options.
+	 * This plugin's settings options.
 	 *
-	 * @since    1.0.0
-	 * @access   protected
+	 * @since  1.0.0
+	 * @access protected
+	 * @var    array $options Array of this plugin settings options.
 	 */
 	protected $options = array();
 
 
 	/**
-	 * Construct.
+	 * Init.
 	 *
-	 * @since    1.0.0
+	 * @since 1.0.0
 	 */
 	public function __construct() {
 
-		$this->plugin_name = 'wp-php-console';
+		$this->plugin_slug = 'WP PHP Console';
+		$this->plugin_slug = 'wp-php-console';
 		$this->version     = '1.4.0';
 		$this->options     = get_option( 'wp_php_console' );
 
@@ -66,8 +79,9 @@ class WP_PHP_Console {
 		// so all temporary data will be stored in $_SESSION.
 		// But there is some problem with frameworks like WordPress that override PHP session handler.
 		// PHP Console has alternative storage drivers for this - we will write to a temporary file:
-		$phpcdir = dirname( __FILE__ ) . '/tmp';
+		$phpcdir  = dirname( __FILE__ ) . '/tmp';
 		$make_dir = wp_mkdir_p( $phpcdir );
+
 		if ( $make_dir === true ) {
 			PhpConsole\Connector::setPostponeStorage(
 				new PhpConsole\Storage\File( $phpcdir . '/' . md5( __FILE__ ) . '_pc.data', false )
@@ -83,7 +97,7 @@ class WP_PHP_Console {
 				// Only if PC not registered yet
 				PhpConsole\Helper::register( );
 			}
-			// PC::debug( 'PC::debug() is available');
+			// PC::debug( 'PC::debug() is available' );
 		}
 
 		// Apply 'stack' option to PHP Console
@@ -109,11 +123,12 @@ class WP_PHP_Console {
 	/**
 	 * Set plugin text domain.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	public function set_locale() {
+
 		load_plugin_textdomain(
-			$this->plugin_name,
+			$this->plugin_slug,
 			false,
 			dirname( plugin_basename( __FILE__ ) ) . '/languages/' );
 	}
@@ -122,15 +137,15 @@ class WP_PHP_Console {
 	/**
 	 * Plugin Settings menu.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	public function register_settings_page() {
 
 		add_options_page(
-			__( 'WP PHP Console', $this->plugin_name ),
-			__( 'WP PHP Console', $this->plugin_name ),
+			__( 'WP PHP Console', 'wp-php-console' ),
+			__( 'WP PHP Console', 'wp-php-console' ),
 			'manage_options',
-			$this->plugin_name,
+			$this->plugin_slug,
 			array( $this, 'settings_page' )
 		);
 
@@ -142,7 +157,7 @@ class WP_PHP_Console {
 	/**
 	 * Register plugin settings.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	function register_settings() {
 
@@ -154,56 +169,56 @@ class WP_PHP_Console {
 
 		add_settings_section(
 			'wp_php_console',
-			__( 'Settings', $this->plugin_name ),
+			__( 'Settings', 'wp-php-console' ),
 			array( $this, 'settings_info' ),
-			$this->plugin_name
+			$this->plugin_slug
 		);
 
 		add_settings_field(
 			'password',
-			__( 'Password', $this->plugin_name ),
+			__( 'Password', 'wp-php-console' ),
 			array( $this, 'password_field' ),
-			$this->plugin_name,
+			$this->plugin_slug,
 			'wp_php_console'
 		);
 
 		add_settings_field(
 			'ssl',
-			__( 'Allow only on SSL', $this->plugin_name ),
+			__( 'Allow only on SSL', 'wp-php-console' ),
 			array( $this, 'ssl_field' ),
-			$this->plugin_name,
+			$this->plugin_slug,
 			'wp_php_console'
 		);
 
 		add_settings_field(
 			'ip',
-			__( 'Allowed IP Masks', $this->plugin_name ),
+			__( 'Allowed IP Masks', 'wp-php-console' ),
 			array( $this, 'ip_field' ),
-			$this->plugin_name,
+			$this->plugin_slug,
 			'wp_php_console'
 		);
 
 		add_settings_field(
 			'register',
-			__( 'Register PC Class ', $this->plugin_name ),
+			__( 'Register PC Class ', 'wp-php-console' ),
 			array( $this, 'register_field' ),
-			$this->plugin_name,
+			$this->plugin_slug,
 			'wp_php_console'
 		);
 
 		add_settings_field(
 			'stack',
-			__( 'Show Call Stack', $this->plugin_name ),
+			__( 'Show Call Stack', 'wp-php-console' ),
 			array( $this, 'stack_field' ),
-			$this->plugin_name,
+			$this->plugin_slug,
 			'wp_php_console'
 		);
 
 		add_settings_field(
 			'short',
-			__( 'Short Path Names', $this->plugin_name ),
+			__( 'Short Path Names', 'wp-php-console' ),
 			array( $this, 'short_field' ),
-			$this->plugin_name,
+			$this->plugin_slug,
 			'wp_php_console'
 		);
 
@@ -213,7 +228,7 @@ class WP_PHP_Console {
 	/**
 	 * Settings Page Password field.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	public function password_field() {
 
@@ -221,9 +236,9 @@ class WP_PHP_Console {
 			'<input type="password" id="wp-php-console-password" name="wp_php_console[password]" value="%s" />',
 			isset( $this->options['password'] ) ? esc_attr( $this->options['password'] ) : ''
 		);
-		echo '<label for="wp-php-console-ip">' . __( 'Required', $this->plugin_name ) . '</label>';
+		echo '<label for="wp-php-console-ip">' . esc_html__( 'Required', 'wp-php-console' ) . '</label>';
 		echo '<br />';
-		echo '<small class="description">' . __( 'The password for eval terminal. If empty, the plugin will not work.', $this->plugin_name ) . '</small>';
+		echo '<small class="description">' . __( 'The password for the eval terminal. If empty, the plugin will not work.', 'wp-php-console' ) . '</small>';
 
 	}
 
@@ -231,7 +246,7 @@ class WP_PHP_Console {
 	/**
 	 * Settings Page SSL option field.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	public function ssl_field() {
 
@@ -241,9 +256,9 @@ class WP_PHP_Console {
 			'<input type="checkbox" id="wp-php-console-ssl" name="wp_php_console[ssl]" value="1" %s /> ',
 			$ssl ? 'checked="checked"' : ''
 		);
-		echo '<label for="wp-php-console-ssl">' . __( 'Yes (optional)', $this->plugin_name ) . '</label>';
+		echo '<label for="wp-php-console-ssl">' . esc_html__( 'Yes', 'wp-php-console' ) . '</label>';
 		echo '<br />';
-		echo '<small class="description">' . __( 'Choose if you want the eval terminal to work only on a SSL connection.', $this->plugin_name ) . '</small>';
+		echo '<small class="description">' . __( 'Tick this option if you want the eval terminal to work only on a SSL connection.', 'wp-php-console' ) . '</small>';
 
 	}
 
@@ -251,7 +266,7 @@ class WP_PHP_Console {
 	/**
 	 * Settings page IP Range field.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	public function ip_field() {
 
@@ -259,9 +274,10 @@ class WP_PHP_Console {
 			'<input type="text" class="regular-text" id="wp-php-console-ip" name="wp_php_console[ip]" value="%s" /> ',
 			isset( $this->options['ip'] ) ? esc_attr( $this->options['ip']) : ''
 		);
-		echo '<label for="wp-php-console-ip">' . __( 'IP addresses (optional)', $this->plugin_name ) . '</label>';
+		echo '<label for="wp-php-console-ip">' . esc_html__( 'IP addresses (optional)', 'wp-php-console' ) . '</label>';
 		echo '<br />';
-		echo '<small class="description">' . __( 'You may specify an IP address (e.g. 192.169.1.50), a range of addresses (192.168.*.*) or multiple addresses, comma separated (192.168.10.25,192.168.10.28) to grant access to eval terminal.', $this->plugin_name ) . '</small>';
+		// translators VVV Varying Vagrant Vagrants default IP address
+		printf ( __( '<small class="description">' . __( 'You may specify an IP address (e.g. <code>192.168.50.4</code>, %s default IP address), a range of addresses (<code>192.168.*.*</code>) or multiple addresses, comma separated (<code>192.168.10.25,192.168.10.28</code>) to grant access to the eval terminal.', 'wp-php-console' ) . '</small>' ), '<a href="https://github.com/Varying-Vagrant-Vagrants/VVV">Varying Vagrant Vagrants</a>' );
 
 	}
 
@@ -269,7 +285,7 @@ class WP_PHP_Console {
 	/**
 	 * Settings page Register PC Class field.
 	 *
-	 * @since   1.2.4
+	 * @since 1.2.4
 	 */
 	public function register_field() {
 
@@ -279,9 +295,9 @@ class WP_PHP_Console {
 			'<input type="checkbox" id="wp-php-console-register" name="wp_php_console[register]" value="1" %s /> ',
 			$register ? 'checked="checked"' : ''
 		);
-		echo '<label for="wp-php-console-register">' . __( 'Yes (optional)', $this->plugin_name ) . '</label>';
+		echo '<label for="wp-php-console-register">' . esc_html__( 'Yes', 'wp-php-console' ) . '</label>';
 		echo '<br />';
-		echo '<small class="description">' . __( 'Choose to register PC in the global namespace. Allows to write PC::debug($var, $tag) or PC::magic_tag($var) instructions in PHP to inspect $var in the JavaScript-console.', $this->plugin_name ) . '</small>';
+		echo '<small class="description">' . __( 'Choose to register PC in the global namespace. Allows to write <code>PC::debug(&#36;var, &#36;tag)</code> or <code>PC::magic_tag(&#36;var)</code> instructions in PHP to inspect <code>&#36;var</code> in the JavaScript console.', 'wp-php-console' ) . '</small>';
 
 	}
 
@@ -289,7 +305,7 @@ class WP_PHP_Console {
 	/**
 	 * Settings page Show Call Stack field.
 	 *
-	 * @since   1.2.4
+	 * @since 1.2.4
 	 */
 	public function stack_field() {
 
@@ -299,9 +315,9 @@ class WP_PHP_Console {
 			'<input type="checkbox" id="wp-php-console-stack" name="wp_php_console[stack]" value="1" %s /> ',
 			$stack ? 'checked="checked"' : ''
 		);
-		echo '<label for="wp-php-console-stack">' . __( 'Yes (optional)', $this->plugin_name ) . '</label>';
+		echo '<label for="wp-php-console-stack">' . esc_html__( 'Yes', 'wp-php-console' ) . '</label>';
 		echo '<br />';
-		echo '<small class="description">' . __( 'Choose to also see the call stack when PHP Console writes to the browser\'s JavaScript-console.', $this->plugin_name ) . '</small>';
+		echo '<small class="description">' . __( "Choose to also see the call stack when PHP Console writes to the browser's JavaScript-console.", 'wp-php-console' ) . '</small>';
 
 	}
 
@@ -309,7 +325,7 @@ class WP_PHP_Console {
 	/**
 	 * Settings page Show Short Paths field.
 	 *
-	 * @since   1.2.4
+	 * @since 1.2.4
 	 */
 	public function short_field() {
 
@@ -319,9 +335,9 @@ class WP_PHP_Console {
 			'<input type="checkbox" id="wp-php-console-short" name="wp_php_console[short]" value="1" %s /> ',
 			$short ? 'checked="checked"' : ''
 		);
-		echo '<label for="wp-php-console-short">' . __( 'Yes (optional)', $this->plugin_name ) . '</label>';
+		echo '<label for="wp-php-console-short">' . esc_html__( 'Yes', 'wp-php-console' ) . '</label>';
 		echo '<br />';
-		echo '<small class="description">' . __( 'Choose to shorten PHP Console error sources and traces paths in browser\'s JavaScript-console. Paths like /server/path/to/document/root/WP/wp-admin/admin.php:31 will be displayed as /W/wp-admin/admin.php:31', $this->plugin_name ) . '</small>';
+		echo '<small class="description">' . __( "Choose to shorten PHP Console error sources and traces paths in browser's JavaScript-console. Paths like <code>/server/path/to/document/root/WP/wp-admin/admin.php:31</code> will be displayed as <code>/W/wp-admin/admin.php:31</code>", 'wp-php-console' ) . '</small>';
 
 	}
 
@@ -329,11 +345,9 @@ class WP_PHP_Console {
 	/**
 	 * Sanitize user input in settings page.
 	 *
-	 * @since   1.0.0
-	 *
-	 * @param   string  $input  user input
-	 *
-	 * @return  array   sanitized inputs
+	 * @since  1.0.0
+	 * @param  string $input user input
+	 * @return array  sanitized inputs
 	 */
 	public function sanitize_field( $input ) {
 
@@ -352,8 +366,8 @@ class WP_PHP_Console {
 			}
 
 			$sanitized_input['register'] = empty( $input['register'] ) ? '' : 1;
-			$sanitized_input['stack'   ] = empty( $input['stack'   ] ) ? '' : 1;
-			$sanitized_input['short'   ] = empty( $input['short'] ) ? '' : 1;
+			$sanitized_input['stack']    = empty( $input['stack'] )    ? '' : 1;
+			$sanitized_input['short']    = empty( $input['short'] )    ? '' : 1;
 
 			return $sanitized_input;
 	}
@@ -362,18 +376,18 @@ class WP_PHP_Console {
 	/**
 	 * Settings page.
 	 *
-	 * @since   1.0.0
+	 * @since 1.0.0
 	 */
 	function settings_page() {
 
 		?>
 		<div class="wrap">
-			<h2><?php _e( 'WP PHP Console', $this->plugin_name ); ?></h2>
+			<h2><?php esc_html_e( 'WP PHP Console', 'wp-php-console' ); ?></h2>
 			<hr />
 			<form method="post" action="options.php">
 				<?php
 				settings_fields( 'wp_php_console' );
-				do_settings_sections( $this->plugin_name );
+				do_settings_sections( $this->plugin_slug );
 				submit_button();
 				?>
 			</form>
@@ -392,13 +406,15 @@ class WP_PHP_Console {
 	public function settings_info() {
 
 		?>
-		<p><?php printf( _x( 'This plugin allows you to use %s within your WordPress installation for testing, debugging and development purposes.<br/>Usage instructions:', 'PHP Console', $this->plugin_name ), '<a href="https://github.com/barbushin/php-console" target="_blank">PHP Console</a>' ); ?></p>
+		<p><?php // Translators: %s refers to 'PHP Console' Chrome extension, will print download link for the Chrome extension
+			printf( _x( 'This plugin allows you to use %s within your WordPress installation for testing, debugging and development purposes.<br/>Usage instructions:', 'PHP Console, the Chrome Extension', 'wp-php-console' ), '<a href="https://github.com/barbushin/php-console" target="_blank">PHP Console</a>' ); ?></p>
 		<ol>
-			<li><?php printf( __( 'Make sure you have downloaded and installed %s.', $this->plugin_name ), '<a target="_blank" href="https://chrome.google.com/webstore/detail/php-console/nfhmhhlpfleoednkpnnnkolmclajemef">PHP Console extension for Google Chrome</a>' ); ?></li>
-			<li><?php _e( 'Set a password for the eval terminal in the options below and hit `save changes`.', $this->plugin_name ); ?></li>
-			<li><?php _e( 'Reload any page of your installation and click on the key icon in your Chrome browser address bar, enter your password and access the terminal.', $this->plugin_name ); ?></li>
-			<li><?php _e( 'From the eval terminal you can execute any PHP or WordPress specific function, including functions from your plugins and active theme.', $this->plugin_name ); ?></li>
-			<li><?php _e( 'In your PHP code, you can call PHP Console debug statements like debug($var, $tag) to display PHP variables in the browser\'s JavaScript-console (Ctrl+Shift+J) and optionally filter selected tags through the browser\'s Remote PHP Eval Terminal screen\'s Ignore Debug options.', $this->plugin_name ); ?></li>
+			<li><?php // Translators: Install PHP Console extension for Google Chrome download link
+				printf( _x( 'Make sure you have downloaded and installed %s.', 'PHP Console, the Chrome Extension', 'wp-php-console' ), '<a target="_blank" href="https://chrome.google.com/webstore/detail/php-console/nfhmhhlpfleoednkpnnnkolmclajemef">PHP Console extension for Google Chrome</a>' ); ?></li>
+			<li><?php _e( 'Set a password for the eval terminal in the options below and hit <code>save changes</code>.', 'wp-php-console' ); ?></li>
+			<li><?php _e( 'Reload any page of your installation and click on the key icon in your Chrome browser address bar, enter your password and access the terminal.', 'wp-php-console' ); ?></li>
+			<li><?php _e( 'From the eval terminal you can execute any PHP or WordPress specific function, including functions from your plugins and active theme.', 'wp-php-console' ); ?></li>
+			<li><?php _e( "In your PHP code, you can call PHP Console debug statements like <code>debug(&#36;var, &#36;tag)</code> to display PHP variables in the browser's JavaScript-console (<code>Ctrl Shift J </code>) and optionally filter selected tags through the browser's Remote PHP Eval Terminal screen's Ignore Debug options.", 'wp-php-console' ); ?></li>
 		</ol>
 		<?php
 
@@ -412,11 +428,8 @@ class WP_PHP_Console {
 	 */
 	public function init() {
 
-		$options = $this->options;
-
-		$password = isset( $options['password'] ) ? $options['password'] : '';
-
-		// Display admin notice and abort if no password
+		// Display admin notice and abort if no password has been set
+		$password = isset( $this->options['password'] ) ? $this->options['password'] : '';
 		if ( ! $password ) {
 			add_action( 'admin_notices', array( $this, 'password_notice' ) );
 			return; // abort
@@ -430,17 +443,20 @@ class WP_PHP_Console {
 		$connector = PhpConsole\Connector::getInstance();
 		$connector->setPassword( $password );
 
+		// PhpConsole instance
 		$handler = PhpConsole\Handler::getInstance();
 		if ( PhpConsole\Handler::getInstance()->isStarted() != true ) {
 			$handler->start();
 		}
 
-		$enableSslOnlyMode = isset( $options['ssl'] ) ? ( ! empty( $options['ssl'] ) ? $options['ssl'] : '' ) : '';
-		if ( $enableSslOnlyMode == true ) {
+		// Enable SSL-only mode
+		$enableSslOnlyMode = isset( $this->options['ssl'] ) ? ( ! empty( $this->options['ssl'] ) ? $this->options['ssl'] : '' ) : '';
+		if ( $enableSslOnlyMode ) {
 			$connector->enableSslOnlyMode();
 		}
 
-		$allowedIpMasks = isset( $options['ip'] ) ? ( ! empty( $options['ip'] ) ? explode( ',', $options['ip'] ) : '' ) : '';
+		// Restrict IP addresses
+		$allowedIpMasks = isset( $this->options['ip'] ) ? ( ! empty( $this->options['ip'] ) ? explode( ',', $this->options['ip'] ) : '' ) : '';
 		if ( is_array( $allowedIpMasks ) && ! empty( $allowedIpMasks ) ) {
 			$connector->setAllowedIpMasks( (array) $allowedIpMasks );
 		}
@@ -468,11 +484,11 @@ class WP_PHP_Console {
 	 */
 	public function password_notice() {
 
-		$link = admin_url( 'options-general.php?page=wp-php-console' );
+		$settings_page = esc_url( admin_url( 'options-general.php?page=wp-php-console' ) );
 
 		?>
 		<div class="update-nag">
-			<?php printf( __( 'WP PHP Console: remember to %1$s set a password %2$s if you want to enable terminal.', $this->plugin_name ), '<a href="' . $link .'">', '</a>' ); ?>
+			<?php printf( $this->plugin_name . ': ' . __( 'Please remember to %1$s set a password %2$s if you want to enable terminal.', 'wp-php-console' ), '<a href="' . $settings_page .'">', '</a>' ); ?>
 		</div>
 		<?php
 
@@ -480,5 +496,3 @@ class WP_PHP_Console {
 
 
 }
-
-new WP_PHP_Console();